requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.32.0
python-dotenv==1.0.1
aiosmtplib==2.0.0
whisperx==3.2.0
faster-whisper==0.10.0
pyannote.audio==3.1.1
numpy==1.26.4
ffmpeg-python==0.2.0
torch==2.5.0
langchain==0.2.0
langchain-core==0.2.38
langchain-openai==0.1.17
langchain-text-splitters==0.2.4
pydantic==1.10.13